requirements.txt
google-genai
python-dotenv
win10toast
pypdf
